refactor dimension typing

+++ src/config/adwords/reports/ADGROUP_PERFORMANCE_REPORT/allconversions.php
@@ -3,7 +3,7 @@ return [
     "id" => "allconversions",
     "property" => "AllConversions",
     "is_filter" => TRUE,
-    "type" => "double",
+    "type" => "decimal",
     "is_metric" => TRUE,
     "is_dimension" => FALSE,
     "is_percentage" => FALSE
