+++ src/config/adwords/reports/ADGROUP_PERFORMANCE_REPORT/mobilebidmodifier.php
@@ -3,7 +3,7 @@ return [
     "id" => "mobilebidmodifier",
     "property" => "AdGroupMobileBidModifier",
     "is_filter" => TRUE,
-    "type" => "double",
+    "type" => "decimal",
     "is_metric" => FALSE,
     "is_dimension" => TRUE,
     "is_percentage" => FALSE
