+++ src/config/adwords/reports/CAMPAIGN_PERFORMANCE_REPORT/invalidclicks.php
@@ -3,7 +3,7 @@ return [
     "id" => "invalidclicks",
     "property" => "InvalidClicks",
     "is_filter" => TRUE,
-    "type" => "long",
+    "type" => "integer",
     "is_metric" => TRUE,
     "is_dimension" => FALSE,
     "is_percentage" => FALSE
